requirements.txt.txt
Python==3.13.5
pip==25.1.1
Pillow==11.3.0